Integrate React Router for page navigation
React Router was introduced into the project to assist with page navigation. An instance of BrowserRouter is now wrapping the main App component and 'react-router-dom' was added to the package dependencies. The '<Route>' component from 'react-router-dom' is then used to specify the render paths of '/dialogs' and '/profile' which would render the Dialog and Profile components respectively. Links in the Navbar component were also updated to point to correct routes. This change provides seamless navigation between different sections of the application, and it was necessary because previously all components were rendered simultaneously or hardcoded, which is not a good practice for a big application. Furthermore, removed several CSS styles and files that were unused or unnecessary for the current project structure. ReportWebVitals, which is used for performance report, was also removed from the index.js file as it was redundant in this context.