Update routing and application structure
Switched from using ReactDOM.render to ReactDOM.createRoot to enable concurrent mode. Additionally, updated the routing from using BrowserRouter to Routes for a more modern and cleaner way of routing. Also, adjusted the the href for 'Dialogs' in the navbar component. Imported index.css and reportWebVitals for evaluating performance. These changes helps to enhance the app flexibility and performance.